+++ lib/mergometer/reports/individual_review_report.rb
@@ -8,12 +8,12 @@ module Mergometer
         end
 
         def table_keys
-          @table_keys ||= grouped_prs_by_time_and_reviewer.keys.map(&:to_date).map { |d| d.strftime("%Y-%m-%d") }
+          @table_keys ||= grouped_prs_by_time_and_reviewer.keys.map { |d| d.strftime("%Y-%m-%d") }
         end
 
         def data_sets
-          @_data_sets ||= grouped_prs_by_reviewer.map do |k, v|
-            [k, v.map(&:count)]
+          @data_sets ||= grouped_prs_by_time_and_reviewer.values.first.keys.map do |user|
+            [user, grouped_prs_by_time_and_reviewer.values.map { |grouped| grouped[user].count }]
           end.to_h
         end
