+++ lib/mergometer/reports/pr_trend_report.rb
@@ -16,7 +16,7 @@ module Mergometer
         def data_sets
           @data_sets ||= METRICS.map do |metric|
             ["average_#{metric}", grouped_prs_by_time.values.map do |prs|
-              PrReportEntry.new(prs).public_send("average_#{metric}")
+              PullRequests.new(prs).public_send("average_#{metric}")
             end]
           end.to_h
         end
