+++ lib/mergometer/reports/ranking_report.rb
@@ -11,7 +11,7 @@ module Mergometer
         end
 
         def table_keys
-          @table_keys ||= grouped_prs_by_users.keys
+          @table_keys ||= grouped_prs_by_user.keys
         end
 
         def data_sets
@@ -29,8 +29,8 @@ module Mergometer
         end
 
         def all_rankings
-          @all_rankings ||= grouped_prs_by_users.map do |user, user_prs|
-            [user, PrReportEntry.new(user_prs)]
+          @all_rankings ||= grouped_prs_by_user.map do |user, user_prs|
+            [user, PullRequests.new(user_prs)]
           end.to_h
         end
